refactoring: move classes to new package 'node'
